Move some old param_spaces to 'param_spaces_old.py'
